--- para_o_lar/Instrucoes.md
@@ -1,41 +1,0 @@
-# Calma! É só mais uma forma de reforçar o que aprendemos!
-
-Já treinamos com nossos exemplos em aula
-
-## Nosso projetinho
-
-Vamos continuar com o projeto da redes sociais, vamos deixa-los mais robusto. Nao tenha medo de usar tudo o que aprendeu ate agora (lembre-se nada de ruim pode acontecer, esse é momento de testar, treinar e errar)
-
----
-
-### Entendendo o que fazer
-
-Vamos entender melhor os nossos dados e contrato definido para nossa API. 
-O json com os dados que vamos usar na nossa API é o arquivo `estabelecimento.json` que está na pasta.
-
-
-** Vamos cadastrar novos estabelecimentos usando o verbo correto (qual sera o verbo?)
-** Vamos atualizar apenas o endereço de um estabelecimento 
-
-Vamos deixar mais completo a nossa rede social
-
-** Vamos Incluir a opção de dar like no estabelecimento (sim voce consegue fazer isso)
-** Mas e se voce foi no estabelecimento e não gostou? Então vamos tambem criar a opção de deslike.
-
-** Agora a parte criativa fica com você, de um nome para a nossa rede social, voce tem liberdade para criar outras melhorias.
-
-## Como Fazer?
-
-Não se desespere, siga os passos que você fez nas últimas aulas. Estruture seu projeto com a arquitetura MVC: crie o server, as rotas, o app e dentro da pasta SRC insira as pastas e arquivos model e controller. 
-
-
-## Postman
-
-Lembre de usar o postman para testar se a solução tá certinha a partir das rotas criadas.
-
-
-
-## E lembrem...
-Usem o plantão de dúvidas, assistam a gravação da aula. E se ainda assim não conseguir, solicite ajuda da coleguinha. Sabemos que está ficando mais puxado, vamos nos ajudar mutuamente!
-
-
