+++ para_o_lar/api-grab-a-bite/server.js
@@ -1,0 +1,6 @@
+const app = require('./src/app.js');
+const PORT = 7050;
+
+app.listen(PORT, () => {
+    console.log(`Servidor Rodando em http://localhost:${PORT}`);
+})
