+++ para_o_lar/api-grab-a-bite/src/app.js
@@ -1,4 +1,4 @@
-const routesEstabelecimento = require('./routes/rotasEstabelecimentos')
+const routes = require('./routes/estabelecimentosRoutes')
 
 const express = require('express');
 const cors = require('cors');
@@ -8,6 +8,6 @@ const app = express();
 app.use(express.json());
 app.use(cors());
 
-app.use('/estabelecimentos', routesEstabelecimento)
+app.use('/estabelecimentos', routes)
 
 module.exports = app;
