+++ para_o_lar/api-grab-a-bite/src/models/estabelecimentos.json
@@ -1,0 +1,185 @@
+
+[
+        {
+                "id":1,
+                "likes":0,
+                "dislikes":0,
+                "nome":"Fogo de Chão",
+                "website":"https://fogodechao.com.br/",
+                "categoria":"Brasileiro",
+                "endereco":"440 Bellevue Way NE",
+                "bairro":"Lincoln Square South",
+                "cidade": "Bellevue",
+                "telefone":"425-450-4004",
+                "pagamento":["Cartão","Dinheiro","Digital Wallet"],
+                "delivery":true
+        },
+        {
+                "id":2,
+                "likes":0,
+                "dislikes":0,
+                "nome":"Texas de Brazil",
+                "website":"https://texasdebrazil.com/locations/",
+                "categoria":"Brasileiro",
+                "endereco":"4502 S. Steele Street, Ste. 1340A",
+                "bairro": "South Tacoma",
+                "cidade": "Tacoma",
+                "telefone": "253-465-3230",
+                "pagamento":["Cartão","Dinheiro"],
+                "delivery":false
+        },
+        {
+                "id":3,
+                "likes":0,
+                "dislikes":0,
+                "nome":"Grill from Ipanema",
+                "website":"https://www.seattlegrillfromipanema.com/",
+                "categoria":"Brasileiro",
+                "endereco":"2313 1st Ave",
+                "bairro": "Belltown",
+                "cidade": "Seattle",
+                "telefone": "206-457-4885",
+                "pagamento":["Cartão","Dinheiro","Digital Wallet"],
+                "delivery":false
+        },
+        {
+                "id":4,
+                "likes":0,
+                "dislikes":0,
+                "nome":"Botteco Brazil",
+                "website":"https://www.bottecobrazil.com/",
+                "categoria":"Brasileiro",
+                "endereco":"14561 Bothell Way NE",
+                "bairro":"Shoreline",
+                "cidade": "Seattle",
+                "telefone":"206-466-2445",
+                "pagamento":["Cartão","Dinheiro","Digital Wallet"],
+                "delivery":true
+        },
+        {
+                "id":5,
+                "likes":0,
+                "dislikes":0,
+                "nome":"Kizuki",
+                "website":"https://www.kizuki.com/",
+                "categoria":"Asiático",
+                "endereco":"320 E Pine St",
+                "bairro":"Capitol Hill",
+                "cidade": "Seattle",
+                "telefone":"(425) 553-0867",
+                "pagamento":["Cartão", "Dinheiro"],
+                "delivery":true
+        },
+        {
+                "id":6,
+                "likes":0,
+                "dislikes":0,
+                "nome":"Dough Zone",
+                "website":"https://texasdebrazil.com/locations/",
+                "categoria":"Asiático",
+                "endereco":"504 5th Ave S Ste 109",
+                "bairro": "International District",
+                "cidade": "Seattle",
+                "telefone": "(206)285-9999",
+                "pagamento":["Cartão","Dinheiro","Digital Wallet"],
+                "delivery":false     
+        },
+        {
+                "id":7,
+                "likes":0,
+                "dislikes":0,
+                "nome":"LaCaBar",
+                "website":"https://lacabartacoma.com/",
+                "categoria":"Asiático",
+                "endereco":"606 S Junett St",
+                "bairro": "Central Tacoma",
+                "cidade": "Seattle",
+                "telefone": "253 267-5540",
+                "pagamento":["Cartão","Dinheiro","Digital Wallet"],
+                "delivery":true  
+        },
+        {
+                "id":8,
+                "likes":0,
+                "dislikes":0,
+                "nome":"Moshi Moshi",
+                "website":"https://www.moshiramenbar.com/",
+                "categoria":"Asiático",
+                "endereco":"110 Suite B N Tacoma Ave",
+                "bairro": "Stadium District",
+                "cidade": "Tacoma",
+                "telefone": "(253) 301-4688",
+                "pagamento":["Cartão","Dinheiro"],
+                "delivery":false  
+        },
+        {
+                "id":9,
+                "likes":0,
+                "dislikes":0,
+                "nome":"Wooden City",
+                "website":"https://www.woodencitytacoma.com/",
+                "categoria":"Americano",
+                "endereco":"714 Pacific Ave",
+                "cidade": "Seattle",
+                "telefone":"(253) 503-0762",
+                "pagamento":["Cartão","Dinheiro","Digital Wallet"],
+                "delivery":false
+        },
+        {
+                "id":10,
+                "likes":0,
+                "dislikes":0,
+                "nome":"Dick’s Drive-In",
+                "website":"https://www.ddir.com/",
+                "categoria":"Americano",
+                "endereco":"115 Broadway E",
+                "bairro": "Capitol Hill",
+                "cidade": "Seattle",
+                "telefone":"206-323-1300",
+                "pagamento":["Cartão","Dinheiro","Digital Wallet"],
+                "delivery":false
+        },
+        {
+                "id":11,
+                "likes":0,
+                "dislikes":0,
+                "nome":"Shelter Lounge",
+                "website":"https://www.shelterlounge.com/",
+                "categoria":"Vegetariano",
+                "endereco":"7110 E Green Lake Drive N",
+                "bairro": "Green Lake",
+                "cidade": "Seattle",
+                "telefone":"(206) 420-7452",
+                "pagamento":["Cartão","Dinheiro","Digital Wallet"],
+                "delivery":true
+
+        },
+        {
+                "id":12,
+                "likes":0,
+                "dislikes":0,
+                "nome":"Cycledogs" ,
+                "website":"https://www.ddir.com/",
+                "categoria":"Vegetariano",
+                "endereco":"5410 17th Ave NW",
+                "bairro": "Ballard",
+                "cidade": "Seattle",
+                "telefone":"206-323-1300",
+                "pagamento":["Cartão","Dinheiro","Digital Wallet"],
+                "delivery":false
+        },
+        {
+                "id":13,
+                "likes":0,
+                "dislikes":0,
+                "nome":"El Chupacabra" ,
+                "website":"http://www.elchupacabraalki.com/",
+                "categoria":"Mexicana",
+                "endereco":"2620 Alki Ave SW",
+                "bairro": "Alki",
+                "cidade": "Seattle",
+                "telefone":"206-933-7344",
+                "pagamento":["Cartão","Dinheiro","Digital Wallet"],
+                "delivery":false
+        }
+]
