+++ para_o_lar/api-grab-a-bite/src/routes/estabelecimentosRoutes.js
@@ -4,10 +4,11 @@ const express = require('express');
 
 const router = express.Router();
 
-router.post('/criar', controller.createEstabelecimento)
+router.post('/', controller.createEstabelecimento)
 router.get('/', controller.getAll)
 router.get('/:id', controller.getById)
-router.patch('/likes/:id', controller.likeDislike)
-router.delete('/delete/:id', controller.deleteEstabelecimento)
+router.put('/atualizar/:id', controller.atualizacao)
+router.patch('/:id', controller.likeDislike)
+router.delete('/:id', controller.deleteEstabelecimento)
 
 module.exports = router;
